added security update date check
Since security updates are cumulative, the original patch outlined in $regex won't be present in the results of Get-HotFix. We can simply check the install date of the update since newer security updates will include the PrintNightmare patch.

One shortcoming this might have is a scenario where somebody cleans a PC and installs pre-PrintNightmare security updates (without the patch). This would show the patch as missing but the date check would see that it was installed after the patch's release, and mark is as valid.